Snapshot ee98b298267d0e09d2cd2f0731b6480a56dd48e7 from master branch of git://git.jetbrains.org/idea/community.git

+++ platform/platform-impl/src/com/intellij/openapi/vfs/impl/local/FileWatcher.java
@@ -173,7 +173,7 @@ public class FileWatcher {
   private static boolean isUpToDate(File executable) {
     long length = SystemInfo.isWindows ? 70216 :
                   SystemInfo.isMac ? 13924 :
-                  SystemInfo.isLinux ? SystemInfo.isAMD64 ? 29227 : 22734 :
+                  SystemInfo.isLinux ? SystemInfo.isAMD64 ? 29269 : 22768 :
                   -1;
     return length < 0 || length == executable.length();
   }
@@ -211,7 +211,9 @@ public class FileWatcher {
     }
 
     LOG.info("Starting file watcher: " + myExecutable);
-    final Process process = Runtime.getRuntime().exec(new String[]{myExecutable.getAbsolutePath()});  // use array to allow spaces in path
+    ProcessBuilder processBuilder = new ProcessBuilder(myExecutable.getAbsolutePath());
+    processBuilder.redirectErrorStream(true);
+    final Process process = processBuilder.start();
     myProcessHandler = new MyProcessHandler(process);
     myProcessHandler.addProcessListener(new MyProcessAdapter());
     myProcessHandler.startNotify();
@@ -229,11 +231,18 @@ public class FileWatcher {
     final OSProcessHandler processHandler = myProcessHandler;
     if (processHandler != null) {
       if (!processHandler.isProcessTerminated()) {
+        boolean forceQuite = true;
         try {
           writeLine(EXIT_COMMAND);
+          forceQuite = !processHandler.waitFor(500);
+          if (forceQuite) {
+            LOG.warn("File watcher is still alive. Doing a force quit.");
+          }
         }
         catch (IOException ignore) { }
-        processHandler.destroyProcess();
+        if (forceQuite) {
+          processHandler.destroyProcess();
+        }
       }
 
       myProcessHandler = null;
@@ -343,19 +352,24 @@ public class FileWatcher {
     protected boolean useAdaptiveSleepingPolicyWhenReadingOutput() {
       return true;
     }
+
+    @Override
+    protected boolean processHasSeparateErrorStream() {
+      return false;
+    }
   }
 
   public boolean isWatched(@NotNull final VirtualFile file) {
     if (isOperational()) {
       synchronized (myLock) {
-        return !checkWatchable(file.getPresentableUrl(), true).isEmpty();
+        return !checkWatchable(file.getPresentableUrl(), true, true).isEmpty();
       }
     }
     return false;
   }
 
   @NotNull
-  private Collection<String> checkWatchable(final String reportedPath, final boolean checkParent) {
+  private Collection<String> checkWatchable(String reportedPath, boolean isExact, boolean fastPath) {
     if (reportedPath == null) return Collections.emptyList();
 
     myAllPaths.clear();
@@ -372,22 +386,31 @@ public class FileWatcher {
     myWatchedPaths.clear();
     ext:
     for (String path : myAllPaths) {
-      for (String root : myRecursiveWatchRoots) {
-        if (FileUtil.startsWith(path, root)) {
-          myWatchedPaths.add(path);
-          continue ext;
-        }
-      }
+      if (fastPath && !myWatchedPaths.isEmpty()) break;
 
       for (String root : myFlatWatchRoots) {
         if (FileUtil.pathsEqual(path, root)) {
           myWatchedPaths.add(path);
           continue ext;
         }
-        if (checkParent) {
-          final File parentFile = new File(path).getParentFile();
-          if (parentFile != null && FileUtil.pathsEqual(parentFile.getPath(), root)) {
+        if (isExact) {
+          String parentPath = new File(path).getParent();
+          if (parentPath != null && FileUtil.pathsEqual(parentPath, root)) {
             myWatchedPaths.add(path);
+            continue ext;
+          }
+        }
+      }
+
+      for (String root : myRecursiveWatchRoots) {
+        if (FileUtil.startsWith(path, root)) {
+          myWatchedPaths.add(path);
+          continue ext;
+        }
+        if (!isExact) {
+          String parentPath = new File(root).getParent();
+          if (parentPath != null && FileUtil.pathsEqual(path, parentPath)) {
+            myWatchedPaths.add(root);
             continue ext;
           }
         }
@@ -523,9 +546,9 @@ public class FileWatcher {
       notifyOnEvent();
     }
 
-    private void processChange(final String path, final WatcherOp op) {
+    private void processChange(String path, WatcherOp op) {
       if (SystemInfo.isWindows && op == WatcherOp.RECDIRTY && path.length() == 3 && Character.isLetter(path.charAt(0))) {
-        final VirtualFile root = LocalFileSystem.getInstance().findFileByPath(path);
+        VirtualFile root = LocalFileSystem.getInstance().findFileByPath(path);
         if (root instanceof NewVirtualFile) {
           ((NewVirtualFile)root).markDirtyRecursively();
         }
@@ -535,8 +558,8 @@ public class FileWatcher {
       }
 
       synchronized (myLock) {
-        final boolean checkParent = !(op == WatcherOp.DIRTY || op == WatcherOp.RECDIRTY);
-        final Collection<String> paths = checkWatchable(path, checkParent);
+        boolean exactPath = op != WatcherOp.DIRTY && op != WatcherOp.RECDIRTY;
+        Collection<String> paths = checkWatchable(path, exactPath, false);
 
         if (paths.isEmpty()) {
           if (LOG.isDebugEnabled()) {
@@ -554,8 +577,8 @@ public class FileWatcher {
           case CREATE:
           case DELETE:
             for (String p : paths) {
-              final File parent = new File(p).getParentFile();
-              myDirtyPaths.add(parent != null ? parent.getPath() : p);
+              String parent = new File(p).getParent();
+              myDirtyPaths.add(parent != null ? parent : p);
             }
             break;
